# Conflicts:
#	electron-app/src/app.ts
#	package.json
#	typings/ipcEvents.ts
#	webapp/package.json
#	webapp/src/app/rootReducer.ts
#	webapp/src/app/update.ipcRenderer.ts
#	webapp/src/constants/configs.ts
#	webapp/src/containers/MasternodesPage/index.tsx
#	webapp/src/containers/MasternodesPage/saga.tsx
#	webapp/src/containers/MasternodesPage/service.tsx
#	webapp/src/containers/Sidebar/index.tsx
#	webapp/src/utils/utility.ts

+++ electron-app/src/utils.ts
@@ -25,10 +25,10 @@ import {
   TESTNET_BASE_FOLDER_REINDEX,
   MAINNET_BASE_FOLDER_REINDEX,
 } from './constants';
-import { DEFAULT_RPC_ALLOW_IP } from '@defi_types/settings';
 import { DAT_FILE_TYPE } from '@defi_types/fileExtensions';
 import * as log from '././services/electronLogger';
 import { IPCResponseModel } from '@defi_types/common';
+import { ADDNODE, MASTERNODE_OPERATOR } from '../../typings/rpcConfig';
 
 export const getPlatform = () => {
   switch (platform()) {
@@ -46,26 +46,10 @@ export const getPlatform = () => {
   }
 };
 
-export const getBinaryParameter = (obj: any = {}) => {
-  let remote: any = {
-    rpcallowip: '',
-    rpcauth: '',
-    rpcport: 0,
-    rpcuser: '',
-    rpcpassword: '',
-    rpcbind: '',
-    datadir: '',
-  };
-  remote.rpcallowip = DEFAULT_RPC_ALLOW_IP;
-  if (!!obj && Array.isArray(obj.remotes)) {
-    remote = Object.assign({}, remote, obj.remotes[0]);
-    remote.rpcbind = obj.remotes[0].rpcconnect;
-    delete remote.rpcconnect;
-  }
-  return Object.keys(remote).map((key) => `-${key}=${remote[key]}`);
-};
-
-export const responseMessage = <T>(success: boolean, res: T): IPCResponseModel<T> => {
+export const responseMessage = <T>(
+  success: boolean,
+  res: T
+): IPCResponseModel<T> => {
   if (success) {
     return { success: true, data: res };
   }
@@ -89,7 +73,22 @@ export const createDir = (dirPath: string) => {
 
 // Get file data
 export const getFileData = (filePath: string, format: string = 'utf-8') => {
-  return fs.readFileSync(filePath, format);
+  const fileData = fs.readFileSync(filePath, format);
+  return formatConfigFileRead(fileData);
+};
+
+// Add squarebrackets masternode_operator in config file
+export const formatConfigFileRead = (fileData: string) => {
+  return fileData
+    .replace(new RegExp(MASTERNODE_OPERATOR, 'gi'), `${MASTERNODE_OPERATOR}[]`)
+    .replace(new RegExp(ADDNODE, 'gi'), `${ADDNODE}[]`);
+};
+
+// Remove squarebrackets masternode_operator in config file
+export const formatConfigFileWrite = (fileData: string) => {
+  return fileData
+    .replace(/masternode_operator[\[\]']+/g, MASTERNODE_OPERATOR)
+    .replace(/addnode[\[\]']+/g, ADDNODE);
 };
 
 // write / append on UI config file
@@ -220,8 +219,10 @@ export const deletePeersFile = () => {
     const baseFolder = getBaseFolderReindex();
     const destFileName = `peers.dat`;
     const destFilePath = path.join(baseFolder, destFileName);
-    deleteFile(destFilePath);
-    log.info(`Deleted peers file in ${destFilePath}`);
+    if (checkPathExists(destFilePath)) {
+      deleteFile(destFilePath);
+      log.info(`Deleted peers file in ${destFilePath}`);
+    }
   } catch (error) {
     log.error(error);
   }
@@ -239,7 +240,9 @@ export const deleteBlocksAndRevFiles = () => {
         (file?.includes(BLK_FILE) || file?.includes(REV_FILE))
       ) {
         log.info(`Deleting ${blkFile}...`);
-        deleteFile(blkFile);
+        if (checkPathExists(blkFile)) {
+          deleteFile(blkFile);
+        }
       }
     });
     log.info('Delete Block and Rev Files completed...');
@@ -253,8 +256,10 @@ export const deleteBanlist = () => {
     const baseFolder = getBaseFolderReindex();
     const destFileName = `banlist.dat`;
     const destFilePath = path.join(baseFolder, destFileName);
-    deleteFile(destFilePath);
-    log.info(`Deleted banlist file in ${destFilePath}`);
+    if (checkPathExists(destFilePath)) {
+      deleteFile(destFilePath);
+      log.info(`Deleted banlist file in ${destFilePath}`);
+    }
   } catch (error) {
     log.error(error);
   }
